rename optimistic_block to optimistic
