+++ crates/astria-sequencer/src/app/mod.rs
@@ -38,7 +38,7 @@ use astria_core::{
     },
     sequencerblock::v1::{
         block::SequencerBlock,
-        optimistic_block::SequencerBlockCommit,
+        optimistic::SequencerBlockCommit,
     },
     Protobuf as _,
 };
